Use relative offsets instead of absolute pointers
By using absolute pointers, if the variant object gets copied or moved,
the base pointer will still point to its old memory location leading to
data sharing (in case of copies) or pointing to invalid memory (in case
of move).

In order to avoid this, we now store the offset to where the wrapped
object begins, relative to the memory address of the variant itself.
This metric is then invariant, if the memory location of the variant
itself changes.