Ensure copy/move-assignment operator is used
Before, it could happen (always happened?) that instead of using the
defaulted copy- or move-assignment operator of polymorphic_variant, the
templated assignment operator for generic types was used.
However, the templated version assigns the type to the internally used
variant and if the polymorphic_variant doesn't happen to convert to a
type that the variant can be assigned to from, a compile error would be
emitted.