Disable generic operator= for incompatible types
If one was to use the generic (templated) operator= with a type that
can't actually be assigned to the underlying variant, the compile error
would happen inside the implementation of operator=, which would make it
reasonably hard to determine what the actual error is and where it is
coming from.
By disabling the templated operator= for incompatible types, the compile
error will happen at the position where the incompatible assignment will
happen (telling you that there is no suitable operator= for the
requested type), which should make it much easier to understand what is
going on.