Disable generic constructor for incompatible types
Same as with the previously unrestricted templated operator=, the
templated constructor would also happily accept any type and if the type
didn't happen to be assignable to the underlying variant, a compile
error would be emitted inside the constructor's implementation (throwing
pages upon pages of error messages at the user).
By disabling the constructor for incompatible types, the error is
instead created at the place where the user tries to use the constructor
with the wrong type and the error message will be much more helpful by
telling the user that such a constructor doesn't exist.