Allow implicit construction
The requirement of always having to construct polymorphic_variants
explicitly turned out to be rather bothersome and in general doing no
good. It was kept explicit to ensure that the code is explicit when
converting something to the generally less efficient variant type.
However, it was encountered that in such cases, this conversion is
simply required and therefore it's better if the compiler can perform
the conversion than having the user type out the conversion every time
explicitly (which can lead to a lot of boilerplate code that decreases
readability).