Remove variant-like API
The fact that a variant is used internally is an implementation detail
now and direct ways to access the variant have been removed. Same goes
for the support for treating a pv as if it was a variant.

Finally, the main header can now be included as pv/pv.hpp and the pv
datatype is now also available as pv::pv. Old include path and symbol
name remain available.